Update test_implicit_func for LLVM 16 (#17623)
In LLVM 16, -Wincompatible-function-pointer types has been upgraded to an error
by default. The functional change in this PR is to update test_implicit_func
to incorporate this. (Currently it accepts either an error or a warning to avoid a complex DEPS roll).
I also minimized the test to only test the behavior that is unique to emscripten (namely, the fact that
-Wimplicit-function-declaration is an error in gnu89 mode rather than a warning as it is upstream).