Bump min browser versions such they all support llvm default features
LLVM and binaryen updated their defaults to include sign extension and
mutable globals: https://reviews.llvm.org/D125728.

According to https://webassembly.org/roadmap/ the min versions required
for sign extension are:

- firefox: 62
- safai: 14.1
- chrome: 71

And for mutable globals:

- firefox: 61
- safai: ??
- chrome: 72

I believe safari implement mutable globals since the beginning which is
perhaps why there is no min version specified.

Since the minimum firefox and chrome versions are already 65 and 74, the
only version we need to bump is the safari version which needs to go
from 12.0 to 14.1.

Also, drop support for the old EdgeHTML engine by default, we we don't
know the status of these features on that engine (we also don't have any
testing at all of this platform).