Avoid false positive test results from JS backtraces
When running tests against minified JS that entire program can end
up being echoed to stderr because there is only a single line of JS.

In this case the expected string could happen to be part of the program
which would cause a false positive result.

The example of this I ran into was the stack overflow tests checking
for the string "stack overflow".  These tests would always pass on any
kind of crash if the JS was minified.  Even more strange was that the
it would start failing again in EMTEST_VERBOSE mode because we were
using `limit_size` on the return value from run_js (which I don't think
was ever intended).