[Wasm64] Implement emmalloc (#17686)
The primary change here is to use `size_t` in a bunch of places that
previously used `uint32_t`.

This change also increases the alignment of `sbrk()` from 4 to
`alignof(maxalign_t)` (8).

As part of this change I also converted all the tracing/logging in
emmalloc form EM_ASM to `_emscripten_out`/`_emscripten_err`.  The latter
two are disigned for low level tracing an should never allocate heap
memory themselves.  The motivation for this is that `EM_ASM` arguments
that are 64-bit are not yet well supported.  At best the arguments would
come through as BigInts and the EM_ASM code would become more complex,
having to handle both cases.