Always check llvm version and allow warning to be suppressed
Previously we would only check the llvm version if the sanity file was
out-of-date.  However, we want this warning to be shown every time the
compiler is run, and the version information is already retrieved from
clang as part of check_sanity so it doesn't cost anything extra to do it
each time.

I also notices the `-Wno-version-check` flag was not working to suppress
this warning since the check was being done before argument parsing.