[WasmFS] Allow backends to report errors in `close`
This is hard to test because we don't have a way of deterministically making a
close fail in any of our existing backends, although in principle failure can
happen in the node backend. I manually tested that this worked correctly by
synthetically injecting an error on close into the OPFS backend and checking
that it did not cause the program to crash or hang and that errno was correctly
set afterward.

However, the close method on SyncAccessHandle is infallible, so we don't
actually need any error handling. As an additional change, simplify
`wasmfs_opfs_close_access` to not wait for the file to be asynchronously closed.