[WasmFS] Allow backends to report errors on `setSize`
Add a test that exercises the `set_size_file` path in the OPFS backend since it
should not be possible to get an error on the `set_size_access`
codepath. (Although I wouldn't be surprised if it is possible in a way I haven't
thought of.) Also, as a drive-by, fix an error code in `doTruncate` to be
properly negative.