[WasmFS] Report errors from `getSize`
Update the return type of `getSize` from `size_t` to `off_t`, which allows for
negative error codes. `off_t` is more correct anyway, since in 32-bit mode
it is possible for a file size to be greater than the max value of `size_t`.
Also update the argument to `getSize` from `size_t` to `off_t` to match.

I manually verified that this error handling works with an injected error, but I
was not able to create a test that could trigger the error naturally.