[WasmFS] Allow `move` to return specific error codes
It already supporting returning errors via a `bool`, but it is better to be able
to return precise errors. Also catch and return errors in the OPFS backend,
which did not previously do any error handling for moves.

Tested manually via an error injected into the OPFS backend since I don't know a
way to get the `move` method on the FileSystemFileHandle to reliably fail in a
way that wouldn't cause the syscall implementation to bail out earlier.