[WasmFS] Return errors from `getEntries` and `getNumEntries` (#17807)
Make the return type of `getEntries` a variant that can be either a vector of
entries as before or a negative error code. Update the return type of
`getNumEntries` to be a ssize_t so it can return negative error codes as well.
This caps the number of directory entries that can be reported at around two
billion, but that should be sufficient.

I have manually tested via an injected error that the error handling works
correctly for the OPFS backend, although I was not able to construct a test case
for this.