Don't pass `-fno-inline-functions` when `-Oz`/`-Os` are used
We have no reason not to let clang use its own heuristics here.

As a followup we can consider completely removing the `INLINGING_LIMIT`
setting since it no longer serves any purpose other than to set this
clang flag.

This change comes with come code size wins and some losses due to the
fact that some system libraries are compiled with `-Os` or `-Oz`.  The
interesting changes seem to mostly be wins.