Fix return value of _mmap_js
_mmap_js previously returned an inptr_t that was either a pointer on success or
a negative error code on failure. However, all negative results were interpreted
as failure, so successfully allocated memory in the second half of memory would
have been incorrectly freed before being returned to the user. To fix this bug,
have _mmap_js return 0 on success and write the returned pointer in an
out-param.