Check return codes in `mp_run_process`
This codepath is not normally used since once currently needs to
opt into using via EM_PYTHON_MULTIPROCESSING=1.

However, I'm hoping to remove that limitation in #17881.  This
cleaanup/fix should land before #17881 does.

The primary change here is to switch from `subprocessrun.Popen` to
`subprocess.run` which supports the `check=True` argument.  As well
as avoiding the need for the call to `communicate()`