[WasmFS] Fix a bug and reduce contention on `close` (#17873)
We were previously calling the backend close method on files every time one of
their fds was closed, but we should have only been calling close on the files
when the last reference to one of their file descriptors was removed from the
file file table. Since the `dup` family of system calls copies fds by reference
rather than by value, these are not the same thing. Due to this bug, it was
possible for `close` to be called on a file more than `open` had been called on
it, which could put backends into invalid states and cause crashes. Fix the
problem by maintaining an explicit reference count on open file descriptors. Do
not just reuse the existing shared_ptr use_count to avoid accidentally
overcounting due to temporary copies of the shared_ptr that are not in the
table.

While fixing that bug, also take the opportunity to reduce lock contention due
to `close` by releasing the file table lock before calling into the backend to
close a file. `close` can be an expensive operation in the backend and the file
table has to be locked frequently, so this was previously a major source of
contention.