Fix timeout in run_multiple_processes
The first argument to `communicate` is `stdin` so this parameter wasn't
doing what it was intended to do.  This means that when the max number
of processes was met we could always block until the first of them was
finished, instead of checking the entire list ever 200 milliseconds.

For things like building libcxx (where each object file can take a while
to build) this can slow down the build but reducing the parallelism.