Don't use -std=c99 in test_life
Not sure why it's there in the first place, but with this
`benchmark.test_life` fails with:
```
life.c:71:7: error: call to undeclared function 'usleep'; ISO C99 and later do not
support implicit function declarations [-Wimplicit-function-declaration]
      usleep(20000);
```

Without this I don't see any errors. Also removed it in `test_core.py`
because why not.