Proxy known handlers back to the main thread
When running on a pthread, none of the incoming parameters on the
module object are present. Proxy known handlers back to the main
thread if specified.

This simplifies the previous `onAbort` pthread logic for a more
generic solution that also supports the `onExit`, `print` and
`printErr` handlers that can be overridden on the incoming module.

See: #17688.