Fix acorn-optimizer.js when writing large files to stdout
This was broken in #17948 but it only showed up in binaryen because we
don't have enough large tests to trigger the issue.

The problem seem to be when using `fs.writeSync` in combination with
`process.stdout.fd`.  In this case we can get partial writes, but even
adding a loop can fail with EAGAIN.

Work around the issue by calling process.stdout.write directly like
we do in print (like we used to do before #17948).

Fixes: #17960