Remove remaining uses of __import decorator
It looks they are not necessary anymore since #17411.
`__asyncify_state__import` and `__asyncify_data__import` were added
after that when #15893 landed, and `__c_longjmp_import` was a typo in
the first place (there are only one `_` instead of two `_`s before
`import`) so it was likely not detected when other uses were removed.