JSPI - Fix multiple async calls exiting early.
When multiple async calls happened in a row the asyncify code was getting in a
bad state because of the differences in how JSPI/ASYNCIFY=1 handle rewinding
(JSPI doesn't re-call the supsended function).

This separates out the implementations for ASYNCIFY=1 and JSPI so they have some
shared code and then their own implementations for `handleSleep` and
`handleAsync`.

Originally we'd hoped to keep most of the code and assertions the same for the
two asyncify implementations, but I think they're different enough that they
should be separated out.
1) ASYNCIFY=1 and JSPI handle rewind differently on how they resume.
2) With JSPI the engine handles some of the assertions.
3) The JSPI code can be much smaller.