Minor refactor of run_mulitple_processes (#17896)
Use a map to store running processes.  This makes deletion faster and
also avoid creating tuples to add the `processes` array and avoid the
need to two different index spaces (One for commands and another for
running commands).

Also, means we can use `pop()` to get and remove the process form the
map in a single statement.