Put stack first when not optimizing
Also, when not using santizers (which currently rely on using the start
of memory).

This allows STACK_OVERFLOW_CHECK=1 to detect stack overflow conditions
better without relying on the STACK_OVERFLOW_CHECK=2 binaryen pass.

This works because when stack is placed first in memory stack overflow
results in SP dropping below zero which is a runtime error.  We can
then distinguish such runtime errors by looking at the SP value at the
time of the crash.

This change is part of a sequence of the effort to reduce the default
stack size.  The hope here is that we will be able to accurately catch
overflows in debug builds even without the STACK_OVERFLOW_CHECK=2
binaryen pass, thus minimizing the impact of reducing the stack size.

See #14177