Remove checks for Module object. NFC (#18068)
These scripts are all included after Module object is created in at the
to of shell.js.

`--pre-js` and `--post-js` files can/should all assuming the existence
of the Module object.  The existance of code like this can add to the
confusions, leading folks like me to believe the cannot rely on the
existence of `Module`.