Switch to upstream freetype repo
We were previously pointing at a version_1 tag in our custom repo.
This was basically VER-2.6 which a single patch applied.

I can't see anything in that patch that is useful, perhaps this code
didn't previously compile cleanly with emcc but today it seems to work
just fine:

https://github.com/emscripten-ports/FreeType/commit/40a760c963bc2b76575918ff91a515c8474db4e0

This seems like a good intermediate step before we actually do a
freetype upgrade in #18088.