Use Linux-built cache/sysroot for Mac CI tester  (#18086)
This PR makes the CircleCI builders use the sysroot built by the Linux builder on the mac bot, instead of having the mac bot rebuild all the libraries. It does this by attaching the same workspace that the linux testers do, and then removing the Linux-specific binaries and emsdk, and using the mac EMSDK instead.