Revert "Switch to upstream freetype repo (#18095)"
This reverts commit ba3c10bd4d4f39c59a21a8a36f3751ae11cf4107.

This change accidentally removed an emscripten patch to freetype.
See: https://github.com/pyodide/pyodide/pull/3239#event-7808406311.

Sadly we don't have a test for this patch so it wasn't picked up by our
CI.  I will work on adding a test so we can try to get this updated
again in the future, but for now a quick revert makes the most sense.