[WasmFS] Fix a deadlock on ProxyWorker creation
It was previously possible for a deadlock to occur when a background thread
constructed the OPFS backend then tried to use it. If the main thread was also
trying to do a WasmFS operation at the same time, it could have been blocked on
a lock held by the background thread and therefore would never have received the
request for it to spawn the OPFS dedicated worker. Neither thread can make
progress in this situation and the page will become unresponsive.

To fix the problem in the simplest possible way, wait for the dedicated worker
to start running before returning from the `ProxyWorker` constructor. This
guarantees that the worker will already be running for subsequent OPFS backend
operations, so they will not be blocked from making progress.

Unfortunately this fix also means that trying to create a `ProxyWorker` from the
main thread will deadlock, so in particular the OPFS backend can no longer be
constructed on the main thread. This restriction should be fine for now because
in practice many applications are using -sPROXY_TO_PTHREAD and are already
creating their OPFS backends on background threads.