[WasmFS] Add virtual backend utilities
Add generic subclasses of DataFile, Directory, and Symlink that wrap other files
and pass operations through to them. Use the new utilities in a rewrite of the
ignore-case virtual backend that cleanly separates the underlying "real" file
system from the virtual file system. This also fixes outstanding issues with the
ignore case backend in which the original, non-normalized file names were not
preserved.