Improve & simplify time sync in pthreads
This supersedes solution from #6084 / #10137 with `performance.timeOrigin`.

This property provides the high resolution baseline for `performance.now()` results, which is exactly what we need. Unlike custom `__performance_now_clock_drift` sent as a message to a worker, it won't suffer from the small time difference incurred by `postMessage` itself, making synchronisation more precise, while at the same time allows to remove some custom code, making it a bit simpler too.

I checked caniuse and it's supported in all browsers where SAB are supported, so should be always safe to use when pthreads are enabled.

Note that this changes the absolute value returned from emscripten_get_now, but this is fine because that value is already different across engines (e.g. it might already return Date.now() in some browser), and it's explicitly documented as "is only meaningful in comparison to other calls to this function".

I verified that test_pthread_reltime still passes too.