[emsymbolizer] Restrict choices of source
Currently `-s`/`--source` option can take any string, but only `dwarf`
and `sourcemap` are meaningful values for that. This restricts the
choices to those two values in `argparse`. This also has a benefit of
showing the possible values in the help message, so you can know what to
add after `-s` without checking the source code.