Extract a worker var in allocateUnusedWorker. NFC
This is a minor nit change, but one I keep coming to too often when working on library_pthread.js. It's convenient to have a single variable that you can do extra operations on before/after adding to the pool, and it's also a bit cleaner to avoid the `PThread.unusedWorkers.push` duplication.