Add explicit GL dependency to library_html5_webgl
The dependency wrangling here was a workaround for a problem that was
solved in #16405 (see the command that is being deleted).

Pretty much all the functions in this depend either directly or
indirectly on the global GL object, so it makes sense to use autoAddDeps
rather than try to add individual deps.