Replace use of llvm-dis as the guessed location for LLVM root
generate_config tries to find the location of an LLVM installation based on the
presence of llvm-dis. This will likely work to locate a
toolchain-developer-focused or self-built version of LLVM, but not one that
is part of a Linux system package or mac SDK, because those don't include
llvm-dis. This could fail in 2 ways: if there were a wasm-supporting clang
and lld (which are all that's needed from LLVM for many purposes) but no
llvm-dis (which is not included in emsdk); or if there were an LLVM without
wasm support.
We could use wasm-ld instead, which is one truly reuquired component
(and the presence of which would likely indicate a wasm-supporting clang
to go with it). This could still fail in some cases (e.g. sometimes a
tool like objcopy is needed and might not be present) but no cases
I'm aware of that are not already failure modes of the existing code.