Update/fix CheckSizeType cmake check (#18309)
Update the CheckTypeSize files to v3.10.2 (mostly arbitrary version, but
this is what we use in CI).

Local patch to CheckTypeSize is not much smaller.  Instead of executing
the program and using the return value (doesn't work for return values
greater than 255), we simply inject `-oformat=wasm` onto the link
command, and then we can use `strings` on the binary just like the
upstream check.  This is basically a single line patch against upstream
now.

This change also avoids a second issue which is that node failures (non-zero
return codes) are indistinguishable from non-zero return codes from user
code.

Fixes: #18278 #18238 #17268 #18084 #17811