[Proxying] Add `emscripten_proxy_async_with_callback` (#18350)
Add a public function for asynchronously proxying work and receiving a callback
on the proxying thread once that work has finished. In principle we could also
add `emscripten_proxy_async_with_callback_and_ctx` in case the asynchronously
proxied work is itself asynchronous, but leave that as future work.

Another issue is that the tests for the new API must busy wait for the callback
to be enqueued, since there is no way of blocking until work is available in the
queue. Adding such a capability is also left as future work.

Fixes #18349.