Detect GL version for custom preinitialized contexts (#18340)
For debugging purposes, it can be useful to wrap a WebGL context into a custom
object presenting the same interface (e.g. using
https://github.com/KhronosGroup/WebGLDeveloperTools). This can be passed to
Emscripten using preinitializedWebGLContext. However, currently checks for an
instance of WebGL2RenderingContext, and so it will assume that a custom
context wrapper only supports WebGL 1, even if it is in fact a WebGL 2 context.

To address that, we can parse the VERSION parameter instead, which is guaranteed
to start with "WebGL 1.0 " or "WebGL 2.0 " depending on which API the context
supports. This way, the version of a wrapped context is correctly detected.