Allow pthreads on Node.js without a pthread pool (#18305)
Node.js `worker_threads` are different than browser `Worker`s in that they start spawning synchronously without waiting for an event loop tick.

If we can also avoid waiting for the `"loaded"` event from the worker before sending pthread tasks to it, then we can spawn new pthreads "synchronously" (prepare everything within the same event loop tick), and block the current thread, making the behaviour on Node.js a lot closer to native and avoid the need for a pthread pool even without Asyncify or extra helper workers.

That's what I did in this implementation. Instead of waiting for the worker to tell us it's loaded and ready, I'm sending all commands immediately to the worker. The worker accepts the first "load" message, starts initializing the runtime, and meanwhile queues up any further messages such as `"run"`. Once the runtime is ready, it processes the queue.

I could limit those changes only to Node.js, but it's easier to do both together, as it allows to avoid a custom `worker.runPthread` callback, and should be in theory even a bit faster in browsers too (by avoiding the "loaded" roundtrip before running a pthread).

---

Here's a lifecycle summary by using a `pthread_create` call as an example, where we don't have a pthread pool or it's already exhausted.

Before:

1. Main thread creates a Worker via `new Worker` and subscribes to its messages.
2. Main thread sends a message "load" with the Wasm module and some other bits to the Worker.
3. Worker receives the "load" message with the Wasm module, loads relevant JS files, and asynchronously initializes the runtime.
4. Worker initialization is done, it sends a message "loaded" to the main thread.
5. Main thread receives message "loaded".
6. Main thread sends a message "run" to the Worker with a pointer to pthread callback and other relevant info.
7. Worker receives the message "run" and executes the pthread.

After:

1. Main thread creates a Worker via `new Worker` and subscribes to its messages.
2. Main thread sends a message "load" with the Wasm module and some other bits to the Worker.
3. Main thread sends a message "run" to the Worker with a pointer to pthread callback and other relevant info.
4. Worker receives the "load" message with the Wasm module, loads relevant JS files, and asynchronously initializes the runtime.
5. Worker stores all the other incoming messages into a queue (in this example it's just a message "run").
6. Worker initialization is done, it sends a message "loaded" to the main thread.
7. Worker executes all the queued up messages (in this example just a message "run", so it executes the pthread).

What this reordering means in practice in Node.js is that before this change the main thread couldn't block synchronously right after `pthread_create` because it then didn't ever have a chance to run steps 5-7, because, with the event loop blocked, it wouldn't even receive the message "loaded" from the Worker.

After this change, the main thread doesn't wait for that "loaded" message anymore and instead sends the "run" command immediately in the same event loop tick, putting the responsibility onto the Worker to only execute the code after it's fully loaded. This means the main thread can safely block right after the `pthread_create` call, even though the Worker is not ready yet.