[Proxying] Abort if a callback cannot be scheduled
The only way a callback can fail to be scheduled is on OOM, and at the point
where it is scheduled there is no way to gracefully report an error. Since
programs may hang or otherwise behave arbitrarily poorly if the callback is not
run, eagerly abort the program in this case instead.