Don't implicitly exit the main thread in PROXY_TO_PTHREAD mode (#18372)
maybeExit should never exit when its run from the main thread in
PROXY_TO_PTHREAD mode.  In this mode we always wait for the proxied main
thread to signal the exit of the program.

Without this change the last three lines of the test output are never
printed because the whole program exits after running the callback.