[Wasm64] Fix emscripten_set_timeout and friends
Timeout IDs are all numbers and not bigint so we should use `int` to
represent them in C and `i` to represent in signatures.

Testing this stuff is tricky right now because use an older version
node in our CI tests and functions like clearTimeout and setInterval
are not available under d8.

Split out from #18392