Only use futex_wait_busy on the main browser thread
When I ported this code from JS back in #15742 I mistakenly changed
the semantics here.  This change restores the previous behaviour of
using atomic.wait except in on the main browser thread.

The `futex_wait_busy` can clearly only be used a single main thread and
is not appropriate for using in cases where there are other threads in
the system that also don't support `atomic.wait` (such as IIUC audio
worklets).  To support that case we would need a different busy wait
function that didn't depend on some global singleton like
_emscripten_main_thread_futex.