[Proxying] Pass a result to the callback
Allow the function proxied by `emscripten_proxy_async_with_callback` to return a
`void*` result that will be passed as an argument to the callback. Previously
this could be accomplished by allocating space for the result value on the
context argument passed to the proxied function and the callback, but providing
the ability to pass the value directly in the API is more convenient and can
save a separate allocation.

Fixes #18378.