Add test for TypeIDs diverging between shared library and executable
When the shared library is built with symbol visibility hidden
as the default, the type infos for any type is not available
to anything that links to the library, and the library or
executable will have its own copy of the type info, which
is not merged.

The results in emscripten::internal::TypeID having different
TYPEID values depending on which module does the lookup,
which causes trouble for embind's type registry as that's
indexed on the TYPEIDs.

Context: #16711