Minor refactor of jsifier.js. NFC
Use a helper function to read and preproccess and write files.

This also allows us to print start and end comments for each file we
include, which makes it easier to understand the debug output file.

This change does use this helper for files that were previously not
preprocessed but since those file don't currently have in macros this
should a be a no-op and it opens up the possibility of using macros
in them later on.