Reorder the browser tests in Chrome
Results are uploaded after this step, and only the last results are uploaded
because earlier ones are overwritten (this also seems to be the case when
specifying multiple tests com the test/runner command line). It didn't seem
worthwhile to add multiple test and upload steps because there are only
a few tests in emrun and posixtest.browser currently.