Make `emscripten_check_blocking_allowed()` no-op in release builds
This warning should only be raised when linking with `-sASSERTIONS`
or `-sALLOW_BLOCKING_ON_MAIN_THREAD=0`, where the latter setting
would treat this warning with a higher severity (i.e., causing the
program to abort when the main thread is blocked).

Previously, this warning was always raised on web environments when
not linking with `-sMINIMAL_RUNTIME`.

Resolves: #18855.