Remove unneeded `to_nice_ident` helper. NFC (#18516)
I'm not sure when or why the wasm imported needed to be cleaned up link
this in the past but it looks like make a fastcomp thing.  wasm-ld won't
generated imports with these types of names under any circumstances that
I know of.  And even if it did I think mangling them like this would
just make things more confusing and broken.