Add MIN_NODE_VERSION setting
This, along with #18465 (which run tests on the oldest supported
version of node) should pave the way for us to update the emsdk version
node to something a little more modern.

See https://github.com/emscripten-core/emsdk/pull/829

This first thing I do with this setting is use it do disable
`NODEJS_CATCH_REJECTION` by default when we are targeting node 15 and
above.