Rename asmLibraryArg to wasmImports. NFC
This internal name is left over from the asm.js days.  This new name
IMHO makes it more obvious what this symbol refers to and aligns with
other globals such as wasmTable and wasmMemory.

As a followup I hope to rename `asm` to `wasmExports` but that could
be trickier because its more likely folks are relying on that name
externally.